refactor: 패키지 구조 개선
- webserver 의 패키지가 지나치게 깊어지는 문제 해결
