+++ src/main/java/servlet/SignUpSuccessServlet.java
@@ -1,22 +1,21 @@
-package webserver.servlet;
+package servlet;
 
 import webserver.httpserver.HttpRequest;
 import webserver.httpserver.HttpResponse;
 import webserver.httpserver.StatusCode;
 
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 
 import static utils.FileUtils.getFile;
 
-public class StaticResourceServlet implements Servlet {
+public class SignUpSuccessServlet implements Servlet {
     @Override
     public void handle(HttpRequest request, HttpResponse response) throws IOException {
         response.setStatusCode(StatusCode.OK);
-        response.setHeader("Content-Type", request.guessContentType());
+        response.setHeader("Content-Type", "text/html");
 
-        File file = new File("src/main/resources/static" + request.getUri());
+        File file = new File("src/main/resources/static/registration/success.html");
         byte[] readFile = getFile(file);
         response.setBody(readFile);
     }
